# contracts/control-plane.openapi.yaml
...
openapi: 3.1.0
info:
  title: Bard — Workgroup Control Plane API
  version: 0.0.1
  description: >
    Device enrollment, workgroup create/join/membership, manager approvals,
    membership-derived endpoints, and revocation. Implements the join workflow
    proven at Level 0 in trust/ + tests/test_trust_join.py; real crypto (OpenMLS
    + hybrid PQ + hardware attestation) replaces the stubs behind this same API.
servers:
  - url: https://localhost:8080/cp   # proxied under the Router front door
paths:
  /enroll:
    post:
      summary: Register this device — submit identity key + attestation (join step 1-2)
      operationId: enroll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [keyPackage, attestation]
              properties:
                keyPackage: { $ref: "./trust.schema.yaml#/$defs/KeyPackage" }
                attestation: { $ref: "./trust.schema.yaml#/$defs/Attestation" }
                label: { type: string }
      responses:
        "200":
          description: Enrolled (device identity now known to this control plane)
          content:
            application/json:
              schema:
                type: object
                required: [entityId]
                properties:
                  entityId: { $ref: "./trust.schema.yaml#/$defs/EntityId" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
...
  /fleet:
    get:
      summary: Read-only fleet view for the management console (Sprint B5; feature #64 first slice)
      operationId: getFleet
      description: >
        Joins the per-device enrollment records (enrollment.schema.json) with
        agent heartbeat liveness and advertised power profiles into one
        console-ready view. In v2 this is served by the Registry's
        control-plane surface (registry/app.py); it moves behind the Router
        front door with the rest of this API. The view itself is read-only;
        the console mutates through the /devices/{deviceId}/* actions below
        (Sprint B6).
      security: [{ identitySig: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FleetView" }
        "401": { $ref: "#/components/responses/Error" }

  # --- Console manage actions (Sprint B6; feature #64 core) -------------------
  # The v2 management surface the console drives. approve/revoke existed on the
  # Registry since Sprint B2 (shapes frozen in enrollment.schema.json); they are
  # contracted here too so the console's full surface lives in ONE contract.
  # rename and workgroup are new in B6. Every successful action is recorded in
  # the append-only audit log (GET /audit).
  /devices/{deviceId}/approve:
    post:
      summary: Approve a pending device — it joins the fleet (enrollment.schema.json pending -> active)
      operationId: approveDevice
      security: [{ identitySig: [] }]
      parameters: [{ name: deviceId, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: >
            Approved. The per-device secret is disclosed EXACTLY ONCE here
            (enrollment.schema.json ApproveResponse) for the manager to hand
            to the device operator.
          content:
            application/json:
              schema: { $ref: "./enrollment.schema.json#/$defs/ApproveResponse" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /devices/{deviceId}/revoke:
    post:
      summary: Remove a device — it loses fleet access immediately (-> revoked)
      operationId: revokeDevice
      security: [{ identitySig: [] }]
      parameters: [{ name: deviceId, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: Revoked; the device's credential no longer verifies anywhere.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [device]
                properties:
                  device: { $ref: "./enrollment.schema.json#/$defs/DeviceRecord" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /devices/{deviceId}/rename:
    post:
      summary: Rename a device (the human label shown in the console)
      operationId: renameDevice
      security: [{ identitySig: [] }]
      parameters: [{ name: deviceId, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [label]
              properties:
                label: { type: string, minLength: 1 }
      responses:
        "200":
          description: Renamed.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [device]
                properties:
                  device: { $ref: "./enrollment.schema.json#/$defs/DeviceRecord" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /devices/{deviceId}/workgroup:
    post:
      summary: Put a device in a workgroup, or take it out (name null/absent clears)
      operationId: assignWorkgroup
      description: >
        v2-pragmatic assignment: the console sends the GROUP NAME and the
        Registry derives the WorkgroupId deterministically from it
        ("wg_" + base64url(sha256(name)) truncated, matching
        trust.schema.yaml#/$defs/WorkgroupId), so equal names land in the
        same group across devices with no separate workgroup CRUD. Real
        MLS-backed workgroups (epochs, member roles) replace this derivation
        in v3 behind the same endpoint.
      security: [{ identitySig: [] }]
      parameters: [{ name: deviceId, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: [string, "null"]
                  minLength: 1
                  description: Workgroup name; null (or omitted) removes the device from its group.
      responses:
        "200":
          description: Assignment updated; GET /fleet reflects it on the next read.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [device]
                properties:
                  device: { $ref: "./enrollment.schema.json#/$defs/DeviceRecord" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /audit:
    get:
      summary: Management action history — who did what to which device, when (Sprint B6)
      operationId: getAudit
      description: >
        Append-only audit of console/management actions (approve, revoke,
        rename, workgroup assignment, and — since Sprint B8 — plugin
        enable/disable/config). Manager-authed, newest first. Entries are
        never mutated or deleted.
      security: [{ identitySig: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditView" }
        "401": { $ref: "#/components/responses/Error" }

  # --- Plugin manage (Sprint B8; feature #65 complete) ------------------------
  # The console Plugins pane drives these. The catalog is declared by manifests
  # frozen in plugin-manifest.schema.json (the eds-rules book capstone seam —
  # that contract is NOT touched here); this surface adds the control-plane
  # state AROUND those manifests: enable/disable per device or per workgroup,
  # per-target config (validated against the manifest's configSchema before it
  # is stored — fail fast), and reported plugin health. Every mutation is
  # recorded in the audit log. Enablement is desired state in the control
  # plane; the manager component that launches `entry` is later ROADMAP scope.
  /plugins:
    get:
      summary: Plugin catalog + enable state + health, console-ready (Sprint B8)
      operationId: getPlugins
      security: [{ identitySig: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PluginCatalogView" }
        "401": { $ref: "#/components/responses/Error" }
  /plugins/{pluginId}/enable:
    post:
      summary: Turn a plugin on for one device or one workgroup
      operationId: enablePlugin
      description: >
        Scope is per device OR per workgroup (by NAME — the WorkgroupId is
        derived deterministically, same rule as POST /devices/{deviceId}/workgroup).
        An optional config travels with the enable; the EFFECTIVE config
        (provided here, else the target's stored config, else {}) is validated
        against the manifest's configSchema and the enable is refused (400) on
        any violation — a plugin is never enabled with invalid settings.
      security: [{ identitySig: [] }]
      parameters: [{ name: pluginId, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PluginEnableRequest" }
      responses:
        "200":
          description: Enabled; GET /plugins reflects it on the next read.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PluginStatus" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /plugins/{pluginId}/disable:
    post:
      summary: Turn a plugin off for one device or one workgroup
      operationId: disablePlugin
      description: >
        Idempotent destination (like device revoke): disabling an already-off
        target succeeds. The target's stored config is KEPT so re-enabling
        restores the operator's settings.
      security: [{ identitySig: [] }]
      parameters: [{ name: pluginId, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PluginScope" }
      responses:
        "200":
          description: Disabled.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PluginStatus" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  schemas:
    FleetView:
      type: object
      additionalProperties: false
      required: [devices, generatedAt]
      properties:
        devices:
          type: array
          items: { $ref: "#/components/schemas/FleetDevice" }
        generatedAt: { type: string, format: date-time }
...
    AuditView:
      type: object
      additionalProperties: false
      required: [entries, generatedAt]
      description: GET /audit response — management action history, newest first.
      properties:
        entries:
          type: array
          items: { $ref: "#/components/schemas/AuditEntry" }
        generatedAt: { type: string, format: date-time }
...
    PluginEnableRequest:
      allOf:
        - { $ref: "#/components/schemas/PluginScope" }
        - type: object
          properties:
            config:
              type: object
              description: >
                Optional config stored with the enable. The effective config
                (this, else the stored one, else {}) must validate against
                the manifest's configSchema or the enable is refused.
...
    PluginStatus:
      type: object
      additionalProperties: false
      required: [manifest, enabledDevices, enabledWorkgroups, health]
      description: >
        One catalog entry, console-ready: the declared manifest (the frozen
        plugin-manifest.schema.json seam, embedded verbatim) joined with the
        control-plane state around it.
      properties:
        manifest: { $ref: "./plugin-manifest.schema.json" }
        enabledDevices:
          type: array
          items: { type: string, minLength: 1 }
          description: deviceIds this plugin is enabled for.
        enabledWorkgroups:
          type: array
          description: Workgroups this plugin is enabled for (ids per trust.schema.yaml).
          items:
            type: object
            additionalProperties: false
            required: [workgroupId, name]
            properties:
              workgroupId: { $ref: "./trust.schema.yaml#/$defs/WorkgroupId" }
              name: { type: string, minLength: 1 }
        health:
          type: [array, "null"]
          items: { $ref: "#/components/schemas/PluginHealthEntry" }
          description: >
            Reported health per device, or null when the manifest declares no
            healthEndpoint (unmonitored — typical for a pure client plugin).
    PluginCatalogView:
      type: object
      additionalProperties: false
      required: [plugins, generatedAt]
      description: GET /plugins response — the whole catalog with state, console-ready.
      properties:
        plugins:
          type: array
          items: { $ref: "#/components/schemas/PluginStatus" }
        generatedAt: { type: string, format: date-time }
  securitySchemes:
    identitySig:
      type: http
      scheme: bearer
      description: >
        Hybrid-PQ identity signature (Ed25519+ML-DSA-65) over the request, once
        enrolled. During migration the MVP JWT verifier sits behind this same
        interface (ADR-0006 / TRUST_MODEL §10).
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "./trust.schema.yaml#/$defs/Error" }